fixed collective event counter
