+++ src/arcos_gui/dock_widgets.py
@@ -8,3 +8,4 @@ viewer.window.add_plugin_dock_widget(
 
 
 napari.run()
+
